backend/services/rule_engine/requirements.txt
grpcio==1.71.0
grpcio-health-checking==1.63.0
grpcio-tools==1.63.0
protobuf==5.29.0
python-dotenv==1.0.1
prisma==0.15.0
PyYAML==6.0.2
